2026.03 Release
- 당장 못쓰거나 안쓰는 파일은 일단 제거
- Nuvoton Uspi 오류 수정
- util/MultiMeasure ADC 역순시 계산 틀리는 문제 수정
- Nuvoton Flash의 readUniqueId() 함수 추가
- Nuvoton M25x Uart의 보레이트 계산 개선
- Nuvoton M43x UUART 컴파일 오류 문제 수정
- drv/Uart에 enum 명칭 수정
- Nuvoton M25x Uart의 ISR 코드 개선
- Nuvoton M25x의 HXT Gain 설정 컴파일 에러 문제 수정
- drv/Uart에 Parity 설정 추가
- buadrate -> baudrate로 변경
- Nuvoton Mxxx의 CRC 설정을 보완
- Nuvoton Mxxx Uuart class 정의 관련 에러 수정
- Nuvoton Mxxx의 누락된 HXT 필터 설정 추가
- Nuvoton UART 드라이버 리셋 후 재 초기화시 충돌 문제 수정
- Nuvoton Uart 드라이버에 리셋 기능 추가
- Nuvoton M43x UUART 컴파일 오류 문제 수정
- util/AnalogOutput의 최대 값 초과에 대한 문제 수정
- Nuvoton M25x의 I2C, Qspi 지원 코드 수정
- Nuvoton M25x에서 TIMER 지원 안되는 부분과 ADC 주석 내용 수정

+++ inc/drv/Spi.h
@@ -100,7 +100,7 @@ class Spi : public Drv
 	// 
 	// int8_t data
 	//		전송할 데이터 한 바이트를 설정한다.
-	virtual void send(uint8_t data) __attribute__((optimize("-O1"))) = 0;
+	virtual void send(uint32_t data) __attribute__((optimize("-O1"))) = 0;
 	
 	// 여러 데이터를 전송한다.
 	// 수신 데이터는 무시한다.
@@ -120,7 +120,7 @@ class Spi : public Drv
 	//		수신된 바이트를 반환한다.
 	// int8_t data
 	//		전송할 데이터 한 바이트를 설정한다.
-	virtual uint8_t exchange(uint8_t data) __attribute__((optimize("-O1"))) = 0;
+	virtual uint32_t exchange(uint32_t data) __attribute__((optimize("-O1"))) = 0;
 
 	// 여러 데이터를 교환한다.
 	// 여러 바이트를 보내고 보내는 동안 수신된 데이터를 송신 버퍼에 다시 채운다.
